Removed init() from importing pyos
This was dangerous as it implies a database open connection.  Something
like this should not be done on import time.  Also by this point the
user might not have had time to set the historian so it will be the
default (possibly wront) one!

Got rid of old cruft as well.