Re-layed out the library to be like python
Now there is a pyos.os which has low-level methods similar to those of
python's os.  There is also another low-level module: pyos.db which is
home to a lot of the direct database interaction methods.  These are
then built on by pathlib (again modelled on python's) and pyos.fs which
is a higher-level way of interacting with objects and directories.