Using cmd2s CommandSet for decoupling shells cmds
This facilities allows one to write general commands to be used by the
shell _without_ having to have them as member functions of the Cmd
class.  This is also useful for external libraries that wish to interace
with Pyos as they can define their own commands in the same way.