Large speedup to find
In cases where there is no filter criteria on the type of state of an
object the find method will only search the metadata speeding up the
query a lot (over 10x).  In cases where either state of type filters are
present the performance is unchanged, but generally this already narrows
down the possibilities so the query isn't as expensive.