Command piping for pyos shell commands
Command piping now works for internal pyos commands as well as external
shell commands and they can be mixed.  If there is a name collision e.g.
'ls' then the '!' character can always be used to escape.

The only thing that doesn't work is command completion for piped
commands.  As soon as the first '|' is encountered completion will
assume that commands are shell commands rather than pyos ones.