Converted to new filesystem collection
This new approach uses a separate MongoDB collection to store entries,
similar to inodes, that maintain the structure of the filesystem
hierarchy (directory and object) as well as some metadata.

This give us a bit more flexibilty and doesn't tie filesystem
information to the metadata entries of the obejcts.

The filesystem itself is represented as graph edges where the edge
implcitly describes a source (the parent directory) and a target (the
child object or directory).  This approach decouples information,
allowing, for example, an entry to move path without having to update
any document expect for the entry itself.