Started transition to allowing multiple obj paths
Currently an object can only be in one path at a time, but this will be
changed in due time to make it more like the relationship between
hard links and inodes.  There can be multiple hard links to each inode
and when the reference count drops to zero the inode and corresponding
data get removed.