Significant speedups for navigating directories
Tried to make as much of the filesystem API yield where possible, thus
allowing the user (when in interactive mode) to get feedback in the form
of printed partial results as soon as possible.

Also added ls -1 option which, as in linux, will print a single column
of the contents of a directory, printing each result as it finds it
rather than waiting to have all of the entries and then formatting them
in some pretty way.